Created the Item entity, controller and repository
The Item entity has two fields: id and name. The id is autogenerated and the name (e.g. projector, blackboard) is given as an input when an instance is created.
This entity has a one to many relationship with the Equipment entity.
The reason behind creating this entity is to make filtering easy by having all the avaiable entities in one place and being able to access them independently from the rooms.

Implemented features:
- GET enpoint for retrieving all items, item by id, item by name
- POST enpoint to add a new item
- PUT endpoint to update an existing item
- DELETE endpoint to delete and existing item