Modified the room and building server communication classes to enable filtering
The methods that filter the rooms and buildings can now be accessed through the methods getFilteredBuildings and getFilteredRoomsByBuilding
When calling the methods, no attribute is mandatory except the buildingId when calling getFilteredRoomsByBuilding. All the other attributes can be specified as null and will not be taken into account when filtering.