Modified the server communication class for room and building to improve filtering
The method now as parameters:
- the id of the building, but just for the room method
- the capacity
- a list of String object with the name of the filters

The limit has been increased to 6 filters. However, the list can have more than 6 items so the user can add more filters, but not more than 6 will be considered.
The methods are called
- getFilteredRoomsByBuilding(buildingId, capacity, listOfParameters)
- getFilteredBuildings(capacity,  listOfParameters)
The implementation has changed a bit to make it easy to add more parameters if needed. To do so, the count variable needs to be changed and more parameters should be added in the controllers. Further instructions can be found in the code.
Tests have also been adapted to the new implementation