Fixed the friendship feature
The main problem why the friendship feature was not working was that the user controller cannot be accessed by non-admin users
Thus, the getUserByUsername get endpoint was not working so I moved it to the friend controller.
I also added a constraint in the friendship controller so that a frindship will only exist once. If the friendshipp already existed, then the user will receive a 409 conflict response from the server.