Fixed an issue with the updating of custom events in the calendar.
Now custom events are updated using threads so it would appear faster. Also when loading the calendar view I removed the event hander so that it will not add the events to the database again and now I clear the custom entries when populating the calendar more than once when updating it.