Modified the checks in the room reservation controller.
Now normal users are not allowed to reserve a room after this and next week no matter what day it is today. Before the allowed time was always two weeks from the date of today. Now the date is dynamic and is always no later than the end of next week.